Visualizacion de comentarios cutre
